DRY up Docker container build/rebuild jobs using DockerContainerBuilder PORO
- Extract shared Docker container logic into DockerContainerBuilder model
- Reduce BuildDockerContainerJob from ~160 to 11 lines
- Reduce RebuildDockerContainerJob from 193 to 22 lines
- Use Open3 for secure tar extraction (fixes brakeman warnings)
- Maintain same functionality while eliminating code duplication
